Refactor go script
Separate functions from main function; easier to read
